Fixed loading data in more tests

+++ tests/test_scan.py
@@ -46,7 +46,7 @@ def test_scan_non_poiss():
 def test_scan_poiss():
     n = nptfit.NPTF(tag='Test_Poiss')
 
-    fermi_data = np.load('fermi_data/fermidata_counts.npy')
+    fermi_data = np.load('fermi_data/fermidata_counts.npy').astype(int)
     fermi_exposure = np.load('fermi_data/fermidata_exposure.npy')
     n.load_data(fermi_data, fermi_exposure)
